refactor controller to create or re-enable google users
