+++ spring-back-end/pjl-authentication/pjl-authentication-core/pom.xml
@@ -35,6 +35,11 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>org.springframework.security</groupId>
+            <artifactId>spring-security-oauth2-authorization-server</artifactId>
+        </dependency>
         <!--        -->
         <dependency>
             <groupId>com.google.api-client</groupId>
