+++ spring-back-end/pjl-authentication/pjl-authentication-core/src/main/java/com/pjlsoftware/controller/security/UserController.java
@@ -1,19 +1,16 @@
-package com.pjlsoftware.controller;
+package com.pjlsoftware.controller.security;
 
-import com.pjlsoftware.controller.security.ValidateGoogleAuthToken;
 import com.pjlsoftware.entity.User;
 import com.pjlsoftware.projection.UserProjection;
 import com.pjlsoftware.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.provider.authentication.OAuth2AuthenticationDetails;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -36,28 +33,23 @@ public class UserController {
             produces = {"application/json"}
     )
     public ResponseEntity<String> createGoogleUser() {
-        // Get the authentication object from the SecurityContextHolder
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) SecurityContextHolder.getContext().getAuthentication();
+        String bearerToken = jwtAuthenticationToken.getToken().getTokenValue();
 
-        // Extract the bearer token from the authentication object
-        String bearerToken = null;
-        if (authentication != null && authentication.getDetails() instanceof OAuth2AuthenticationDetails) {
-            OAuth2AuthenticationDetails details = (OAuth2AuthenticationDetails) authentication.getDetails();
-            bearerToken = details.getTokenValue();
-        }
-        User user = ValidateGoogleAuthToken.verifyGoogleAuthToken(credential)
+        User user = ValidateGoogleAuthToken.verifyGoogleAuthToken(bearerToken)
                 .orElseThrow(() -> new RuntimeException("Failed to validate JWT."));
 
         Optional<User> isAlreadyUser = userRepository.findByUsername(user.getUsername());
 
         if (isAlreadyUser.isPresent()) {
-            user.setEnabled(true);
+            User existingUser = isAlreadyUser.get();
+            existingUser.setEnabled(true);
+            userRepository.saveAndFlush(existingUser);
             return new ResponseEntity<>("{\"value\": \"Re-enabled existing user\"}", HttpStatus.CREATED);
         } else {
             userRepository.saveAndFlush(user);
             return new ResponseEntity<>("{\"value\": \"Created new user\"}", HttpStatus.CREATED);
         }
-
     }
 
     @RequestMapping(
