+++ spring-back-end/pjl-authentication/pjl-authentication-model/src/main/java/com/pjlsoftware/entity/User.java
@@ -46,7 +46,7 @@ public class User {
         this.firstName = firstName;
         this.lastName = lastName;
         this.username = username;
-        this.enabled = false;
+        this.enabled = true;
     }
 
     //
